registry: add in local (non-collective) handlers
  - Encode node with a local handler
  - Enables transparent callback-like functionality

+++ src/common.h
@@ -11,7 +11,7 @@
 namespace runtime {
 
 using node_t = int16_t;
-using handler_t = int16_t;
+using handler_t = int32_t;
 using envelope_datatype_t = int8_t;
 using event_t = uint64_t;
 using epoch_t = int32_t;
